requirements.txt
pytesseract
pdf2image
pillow
pypdf
pytest
reportlab
